+++ libs/shared/environments/src/lib/environment.type.ts
@@ -3,4 +3,5 @@ export type environmentType = {
   production: boolean;
   env: envs;
   baseUrl: string;
+  version: string;
 };